% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{adt}
\alias{adt}
\title{Shortcuts for a lazy person}
\usage{
adt(x, keep.rownames = FALSE, ...)
}
\arguments{
\item{x}{}

\item{keep.rownames}{}

\item{...}{}
}
\description{
Shortcuts for a lazy person
}
\note{
This is to be used only for console work and not to be used in functions.
}
