% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbindlist.R
\name{cbindlist}
\alias{cbindlist}
\title{Combine a list of data.tables column-wise}
\usage{
cbindlist(list, key, col_names = NULL)
}
\arguments{
\item{list}{}

\item{col_names}{}
}
\value{

}
\description{
Combine a list of data.tables column-wise
}
