% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcast2.R
\name{dcast2}
\alias{dcast2}
\title{Dcast using list columns}
\usage{
dcast2(
  data,
  formula,
  fun.aggregate = NULL,
  sep = "_",
  ...,
  margins = NULL,
  subset = NULL,
  fill = NULL,
  drop = TRUE,
  value.var = data.table:::guess(data),
  verbose = getOption("datatable.verbose")
)
}
\arguments{
\item{data}{}

\item{formula}{}

\item{fun.aggregate}{}

\item{sep}{}

\item{...}{}

\item{margins}{}

\item{subset}{}

\item{fill}{}

\item{drop}{}

\item{value.var}{}

\item{verbose}{}
}
\description{
Dcast using list columns
}
