% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTH.R
\name{DTH}
\alias{DTH}
\title{Function to implement DTH}
\usage{
DTH(dist_mat, ident, center = "centroid", n.perm = 999, ncores = 2)
}
\arguments{
\item{dist_mat}{The distance matrix, usually obtained from dist() or vegdist()}

\item{ident}{numeric group identifiers, input as factors}

\item{center}{either "centroid" or "median"}

\item{n.perm}{number of permutations}

\item{ncores}{number of cores to be used for computation}
}
\value{
A p-value for the DTH test
}
\description{
Function to implement DTH
}
\examples{
n1 = 30
n2 = 20
p = 300
y1 = matrix(rnbinom(n1*p, mu =  4, size = 1), nrow = n1, ncol = p)
y2 = matrix(rnbinom(n2*p, mu = 4, size = 10), nrow = n2, ncol = p)
y = rbind(y1, y2)
ident = rep(1:2, times = c(n1,n2))
D = vegan::vegdist(y, method = "bray")
DTH(D, ident)
}
