% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_stats.R
\name{dth_aov}
\alias{dth_aov}
\title{Anova formatted for use inside data.tables}
\usage{
dth_aov(formula, DT, by)
}
\arguments{
\item{formula}{}

\item{by}{}
}
\value{

}
\description{
Anova formatted for use inside data.tables
}
