% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_as_matrix.R
\name{dth_as_matrix}
\alias{dth_as_matrix}
\title{Convert a data.table to a matrix easily}
\usage{
dth_as_matrix(DT, rownames.col = NULL, set.rownames = TRUE)
}
\arguments{
\item{DT}{data.table}

\item{rownames.col}{character of which column to use for the matrix rownames. (Default=NULL)}

\item{set.rownames}{logical whether to set the rownames from a column within the data.table. (Default=TRUE)}
}
\description{
Convert a data.table to a matrix easily
}
