% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_clean_names.R
\name{dth_clean_names}
\alias{dth_clean_names}
\title{Return R-compatible column names for a `data.table`}
\usage{
dth_clean_names(.data, unique = FALSE)
}
\arguments{
\item{.data}{}

\item{unique}{}
}
\description{
Return R-compatible column names for a `data.table`
}
