% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_cor_test.R
\name{dth_cor}
\alias{dth_cor}
\title{X-Y Correlation and Statistical Tests for multiple categories using data.table}
\usage{
dth_cor(
  DT,
  x,
  y,
  by,
  method = "spearman",
  alternative = "two.sided",
  exact = NULL,
  conf.level = 0.95,
  continuity = FALSE,
  trim = TRUE
)
}
\arguments{
\item{DT}{}

\item{x}{}

\item{y}{}

\item{by}{}

\item{method}{}

\item{alternative}{}

\item{exact}{}

\item{conf.level}{}

\item{continuity}{}

\item{trim}{}
}
\value{

}
\description{
X-Y Correlation and Statistical Tests for multiple categories using data.table
}
\examples{
require(data.table)
DT <- as.data.table(iris)
# uses spearman method by default
dth_cor(DT, x = 'Sepal.Length', y = 'Sepal.Width', by = 'Species')

# set trim = F if you want all the extra information (mostly useless)
dth_cor(DT, x = 'Sepal.Length', y = 'Sepal.Width', by = 'Species',trim = FALSE)

}
