% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_clip.R
\name{dth_read_clip}
\alias{dth_read_clip}
\title{grab data from the clipboard as a data.table}
\usage{
dth_read_clip(...)
}
\arguments{
\item{...}{}
}
\description{
grab data from the clipboard as a data.table
}
\examples{
\dontrun{
library(data.table)
library(dth)

DT <- dth_read_clip()
}

}
