% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dth_subset.R
\name{dth_subset.list}
\alias{dth_subset.list}
\title{Subset list columns within data.tables}
\usage{
dth_subset.list(column, FUN)
}
\arguments{
\item{column}{}

\item{FUN}{}
}
\value{

}
\description{
Subset list columns within data.tables
}
