% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_lg_cols.R
\name{rm_lg_cols}
\alias{rm_lg_cols}
\title{Remove columns with long character data}
\usage{
rm_lg_cols(DT, nchars = 40)
}
\arguments{
\item{DT}{}

\item{nchars}{}
}
\description{
Remove columns with long character data
}
