% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_list.R
\name{subset_list}
\alias{subset_list}
\title{Subset a `list` type column in a `data.table`}
\usage{
subset_list(column, by, type = "any")
}
\arguments{
\item{column}{list column in data.table to subset}

\item{by}{a character or character vector for subsetting. Can also use function test (e.g. is.na)}

\item{type}{return rows that either have 'any' [default] or 'all' of the `by` values. type = 'none' returns rows with none of the by values. 'only' returns rows that only have the by values.}
}
\description{
Subset a `list` type column in a `data.table`
}
