% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_cols.R
\name{unique_cols}
\alias{unique_cols}
\title{Remove columns with only one unique value from data.tables}
\usage{
unique_cols(df, return.DT = TRUE)
}
\arguments{
\item{df}{}

\item{return.DT}{}
}
\value{

}
\description{
Remove columns with only one unique value from data.tables
}
\examples{


}
